Flask will allow GET reqs to map to gendocs
This commit is the first in a series which will deprecate
support for Apache, and convert the autodocumentation project(s)
to use nginx + python/flask.

map_concepts_as_symlinks.py -> map_entries_to_gendocs.py
This change will adjust the current mapping script to output
a cached json object that will map entry points that a flask
application will accept, and route to the proper documentation
location on the file-system.

Signed-off-by: TJ Balon <tbalon@mitre.org>