Handle Java tools defaults
Some Java ontology-interactive tools were observed to make requests of
ontology IRIs with no 'Accept:' header.  The heuristics implemented
before this patch would treat this as a browser request and serve HTML.
The tools would fail parsing the Gendocs HTML as an ontology file, and
make no further retrieval attempt (e.g. by requesting the original URL,
looping through appending ".rdf", ".ttl", ".xml"; nor looping through
media-types).

This patch adds a test to make a determination of whether a request is
coming from a browser or not.  It is not a general-purpose "Is this a
browser?" function.

References:
* https://github.com/casework/ontology.caseontology.org/issues/12

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>