Fix XML entity substitution
The CASE vocabulary.ttl file uses the prefix "vocab", contrary to other
patterns throughout CASE and UCO.  This would typically not be a
problem, except the process used in the documentation site to convert
Turtle to RDF currently has a defect in handling XML entities, and the
workaround didn't catch this atypical namespace prefix.

This patch also manually handles the substitution in Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>